+++ src/index.js
@@ -6,5 +6,5 @@ export { Router, Router as CliRouter } from './router';
 export { Controller, Controller as CliController } from './controller';
 
 import * as Controllers from './controllers';
-export { Help as HelpController } from './controllers/help';
+export { default as HelpController } from './controllers/help';
 export { Controllers, Controllers as CliControllers };