+++ CHANGELOG.md
@@ -10,7 +10,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   * [api] Renaming CliControllers to Controllers
   * [api] Moving export HelpController from CliControllers to HelpController
 
+* Bug
+  * [help] Fixing parse usage for show help
+
 * Enhancements
   * [examples] Updating basic example
+  * [examples] Moving advanced example from README to examples
   * [code] Replacing azk-dev por pure Babel
   * [code] Replacing gulp for npm/yarn tasks